+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Fixed
+* Fixed an issue where importing some files would not correctly process all LDraw commands.
+
 ## 0.4.6 - 2025-01-28
 ### Changed
 * Adjusted procedural normals for grainy slope materials.